requirement.txt
langchain==0.3.0
langgraph==0.2.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-chroma==0.1.0
streamlit==1.40.0
python-dotenv==1.0.1
tavily-python==0.5.0
pypdf==5.0.0